Restructure node headers
